Changed variable names and wrong data types
Adjusted names to be right according to The Norm, following snake_case and data types to follow the examples

+++ ft_bzero.c
@@ -6,7 +6,7 @@
 /*   By: ppontet <ppontet@student.42lyon.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/05 16:33:04 by ppontet           #+#    #+#             */
-/*   Updated: 2024/11/06 15:18:22 by ppontet          ###   ########lyon.fr   */
+/*   Updated: 2024/11/07 14:39:08 by ppontet          ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@
 void	ft_bzero(void *s, size_t n)
 {
 	unsigned char	*new_pointer;
-	unsigned int	index;
+	size_t			index;
 
 	new_pointer = (unsigned char *)s;
 	index = 0;
