+++ ft_memcpy.c
@@ -6,12 +6,13 @@
 /*   By: ppontet <ppontet@student.42lyon.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/06 15:26:57 by ppontet           #+#    #+#             */
-/*   Updated: 2024/11/06 15:35:15 by ppontet          ###   ########lyon.fr   */
+/*   Updated: 2024/11/07 16:26:46 by ppontet          ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <stddef.h>
 
+//@TODO crash ´your memcpy does not behave well with NULL as both params with size´
 /**
  * @brief Set a memory zone with a value, count times
  *
@@ -24,7 +25,7 @@ void	*ft_memcpy(void *destination, const void *source, size_t size)
 {
 	unsigned char	*dest_pointer;
 	unsigned char	*src_pointer;
-	unsigned int	index;
+	size_t			index;
 
 	dest_pointer = (unsigned char *)destination;
 	src_pointer = (unsigned char *)source;
