+++ ft_memset.c
@@ -6,7 +6,7 @@
 /*   By: ppontet <ppontet@student.42lyon.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/05 15:29:02 by ppontet           #+#    #+#             */
-/*   Updated: 2024/11/06 15:18:49 by ppontet          ###   ########lyon.fr   */
+/*   Updated: 2024/11/07 14:39:41 by ppontet          ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,7 +24,7 @@
 void	*ft_memset(void *pointer, int value, size_t count)
 {
 	unsigned char	*new_pointer;
-	unsigned int	index;
+	size_t			index;
 
 	new_pointer = (unsigned char *)pointer;
 	index = 0;
